Sender-Client-Proxy/requirements.txt
aiosmtpd==1.4.2
aiosmtplib==2.0.0
PyYAML==6.0
